+++ CHANGELOG.md
@@ -18,6 +18,13 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/).
 
 ### Security
 
+## v4.160.0 - 2021-08-23
+
+### Changed
+
+- 注文アイテムとしてのPermitからaccountNumberを削除
+- 決済承認アクションを最適化
+
 ## v4.159.0 - 2021-08-23
 
 ### Changed